# swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": {
			"Fn::Sub": "Your Shared Secret API (${Stage})"
		},
		"version": "1.0",
		"description": "API for managing secrets and caretakers for the Your Shared Secret service."
	},
	"host": "api.yoursharedsecret.com",
	"basePath": "/v1",
	"schemes": [
		"https"
	],
	"paths": {
...
		"/secrets/{secret-id}": {
			"get": {
				"operationId": "GetSecret",
				"summary": "Get a secret.",
				"description": "This can be done either by the owner of the secret or the unlocking caretaker assuming that the 1 week quarantine period has passed and the proper dataKeyDigest is provided.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"description": "Unique secret identifier.",
						"name": "secret-id",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"description": "SHA 256 digest of the key used to encrypt the secret. Only needed when retrieving during an unlock. Encoded using Base64 URL safe encoding.",
						"name": "dataKeyDigest",
						"in": "query",
						"type": "string"
					},
					{
						"description": "Caretaker requesting to fetch the secret. Must also prove knowledge of the dataKey.",
						"name": "caretakerId",
						"in": "query",
						"type": "string"
					},
					{
						"name": "x-yoursharedsecret-ownership",
						"description": "Header proving access to either the private key of the secret or the private key corresponding to the current unlocking caretakers public key or the explicitly specified caretakerId in the query string parameter. See separate section for details.",
						"in": "header",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "200 response",
						"schema": {
							"$ref": "#/definitions/SecretResponse"
						}
					},
					"404": {
						"description": "404 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"423": {
						"description": "423 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					}
				},
		        "x-amazon-apigateway-request-validator": "validator",
				"x-amazon-apigateway-integration": {
					"uri": {
						"Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SecretService.Arn}/invocations"
					},
					"passthroughBehavior": "when_no_match",
					"httpMethod": "POST",
					"type": "aws_proxy"
				}
			},
			"put": {
				"operationId": "UpdateSecret",
				"summary": "Update a secret.",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"description": "Unique secret identifier.",
						"name": "secret-id",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "x-yoursharedsecret-ownership",
						"description": "Header proving access to either the private key of the secret. See separate section for details.",
						"in": "header",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "UpdateSecretRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UpdateSecretRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "200 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"400": {
						"description": "400 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"404": {
						"description": "404 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					}
				},
		        "x-amazon-apigateway-request-validator": "validator",
				"x-amazon-apigateway-integration": {
					"uri": {
						"Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SecretService.Arn}/invocations"
					},
					"passthroughBehavior": "when_no_match",
					"httpMethod": "POST",
					"type": "aws_proxy"
				}
			},
			"delete": {
				"operationId": "DeleteSecret",
				"summary": "Delete a secret.",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"description": "Unique secret identifier.",
						"name": "secret-id",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "x-yoursharedsecret-ownership",
						"description": "Header proving access to either the private key of the secret. See separate section for details.",
						"in": "header",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "200 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"404": {
						"description": "404 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					}
				},
		        "x-amazon-apigateway-request-validator": "validator",
				"x-amazon-apigateway-integration": {
					"uri": {
						"Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SecretService.Arn}/invocations"
					},
					"passthroughBehavior": "when_no_match",
					"httpMethod": "POST",
					"type": "aws_proxy"
				}
			}
		},
...
	},
	"definitions": {
		"MessageResponse": {
			"type": "object",
			"title": "Free form message response.",
			"properties": {
				"message": {
					"description": "Free form message.",
					"type": "string"
				}
			}
		},
...
		"SecretResponse": {
			"type": "object",
			"title": "Get secret response body.",
			"properties": {
				"publicKey": {
					"description": "Public key for secret. Encoded using Base64 URL safe encoding.",
					"type": "string"
				},
				"unlockPublicKey": {
					"description": "Public key of recipient currently performing an unlock of the secret. Encoded using Base64 URL safe encoding.",
					"type": "string"
				},
				"data": {
					"description": "Encrypted data blob. Encoded using Base64 URL safe encoding. The secret key can be either the private key for the secret or the data key of the secret key. The signing key is always the private key of the secret.",
					"type": "string"
				},
				"publishData": {
					"description": "Data to be used to save preferences during creation of secret. The encryption key is the private key for the secret which is also the signing key. Encoded using Base64 URL safe encoding. This field is only returned to the owner of the secret.",
					"type": "string"
				},
				"payDate": {
					"description": "The time up until which this secret is guaranteed with a refund. Returned as milliseconds since January 1st, 1970 UTC.",
					"type": "number"
				}
			},
			"required": [
				"publicKey",
				"payDate"
			]
		},
		"UpdateSecretRequest": {
			"type": "object",
			"title": "Update secret request body.",
			"properties": {
				"dataKeyDigest": {
					"description": "Double SHA512 digest of the key needed for decrypting the data payload. Encoded using Base64 URL safe encoding.",
					"type": "string"
				},
				"paymentType": {
					"description": "What type of payment token is provided. * USD - Normal payment processor. * BTC - Bitcount transaction. * ETH - Etherium transaction. * CPN - Coupon code. Should only be specified if you wish to make an additional payment for this secret.",
					"type": "string",
					"enum": [
						"USD",
						"BTC",
						"ETH",
						"CPN"
					]
				},
				"paymentToken": {
					"description": "Token for proving payment. If USD should be a JSON object containing the key email with the \"email\", associated with the payment and \"nonce\" with the payment authorization nonce from Square. If BTC or ETH should be a transaction hash on the corresponding block chains that have a committed transaction to the correct address with at least the minimum amount (See website for up to date details of the addresses and amounts).",
					"type": "string"
				},
				"data": {
					"description": "Encrypted data blob. Encoded using Base64 URL safe encoding. The decoded data payload must be less than or equal to 1MB in size. The secret key can be either the private key for the secret or the data key of the secret key. The signing key is always the private key of the secret.",
					"type": "string"
				},
				"publishData": {
					"description": "Data to be used to save preferences during creation of secret. The encryption key is the private key for the secret which is also the signing key. Encoded using Base64 URL safe encoding and limited to 1000 bytes encoded.",
					"type": "string"
				}
			},
			"required": [
				"data"
			]
		},
...
	},
...
}